implemented changes, moved most trial configs to experimenthelper.js, timeline now uses the function from config.js
